pyproject.toml
[project]
name = "travel-planner"
version = "0.1.0"
description = ""
authors = [
    {name = "maks",email = "gmb333x2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "dependency-injector (>=4.49.0,<5.0.0)",
    "httpx (>=0.28.0,<1.0.0)",
    "uvicorn (>=0.34.0,<1.0.0)",
    "loguru (>=0.7.3,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.9"
pytest = "^9.0.2"

[tool.ruff]
line-length = 79
target-version = "py313"
# Skip/exclude directories and files
exclude = [
    ".idea/",
    "venv/",
    ".venv/",
    "docs/",
    "scripts/",
    "temp/",
    "alembic/",
    "app_importer.py",
    "tmp_apps.py",
    "playgraund.py",
]

[tool.ruff.lint]
# pycodestyle, pyflakes, isort, pylint, pyupgrade
ignore = ["E501", "UP046"]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.ruff.lint.per-file-ignores]
"src/api/**" = ["B008"]
